+++ src/front/js/layout.js
@@ -17,7 +17,6 @@ import { SobreNosotros } from "./pages/sobreNosotros"
 import { Menu } from "./pages/menu";
 /*import { Profile } from "./pages/profile";*/
 import { Principal } from "./pages/principal";
-import { View } from "./pages/view";
 import { Success } from "./pages/success";
 import { Failure } from "./pages/failure";
 
@@ -45,7 +44,6 @@ const Layout = () => {
                         <Route element={<SobreNosotros />} path="/sobreNosotros" />
                         <Route element={<Menu />} path="/menu" />
                         <Route element={<Principal />} path="/principal" />
-                        <Route element={<View />} path="/view" />
                         <Route element={<Success />} path="/success" />
                         <Route element={<Failure />} path="/failure" />
